+++ test_vectortiles/test_app/functions.py
@@ -1,0 +1,5 @@
+from django.contrib.gis.db.models.functions import GeomOutputGeoFunc
+
+
+class SimplifyPreserveTopology(GeomOutputGeoFunc):
+    pass